pyproject.toml
[project]
name = "beancount-ui"
version = "0.1.0"
description = "A desktop app for managing beancount ledgers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui[native]>=2.0",
    "beancount>=3.0.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
banking = [
    "pyjwt>=2.12.1",
    "cryptography>=46.0.6",
]

[project.scripts]
beancount-ui = "main:run"
